+++ src/components/Home/Skills/BgParticles.tsx
@@ -1,12 +1,13 @@
-import React, { memo, useCallback, useMemo } from 'react';
+import React, { memo, useEffect, useMemo, useState } from 'react';
 
-import { loadFull } from 'tsparticles';
-import Particles from 'react-particles';
+import { loadFull } from "tsparticles";
+import Particles, { initParticlesEngine } from "@tsparticles/react";
 import generateParticlesConfig from '../../../utils/config/particles';
 
 import { useThemeContext } from '@/hoc/ThemeContextProvider';
 
-import type { Engine } from 'tsparticles-engine';
+import type { Engine } from "@tsparticles/engine";
+
 
 //create a map to change color of particles based on theme (light/dark)
 const themeMap = {
@@ -18,16 +19,30 @@ const themeMap = {
 
 const BgParticles = () => {
   const { theme } = useThemeContext();
+  const [ init, setInit ] = useState(false);
+
+  useEffect(() => {
+    initParticlesEngine(async (engine: Engine) => {
+      await loadFull(engine);
+    }).then(() => {
+      setInit(true);
+    });
+  }, []);
 
   const skillParticles = useMemo(() => {
     return generateParticlesConfig(themeMap[theme]);
   }, [theme]);
 
-  const particlesInit = useCallback(async (engine: Engine) => {
-    await loadFull(engine);
-  }, []);
-
-  return <Particles id="tsparticles" className="absolute inset-0" init={particlesInit} options={skillParticles} />;
+  return (
+    <>
+      {init && 
+        <Particles
+          id="tsparticles"
+          options={skillParticles}
+        />
+      }
+    </>
+  )
 };
 
 const areEqual = () => true;